Scheduler: Corrections to YieldAndBalanceLoad and Yield bombing protection.

+++ src/core/hle/kernel/scheduler.h
@@ -155,7 +155,7 @@ private:
     std::array<Common::MultiLevelQueue<Thread*, THREADPRIO_COUNT>, NUM_CPU_CORES> suggested_queue;
     std::atomic<bool> reselection_pending;
 
-    std::array<u64, NUM_CPU_CORES> preemption_priorities = {59, 59, 59, 62};
+    std::array<u32, NUM_CPU_CORES> preemption_priorities = {59, 59, 59, 62};
 
     /// Lists all thread ids that aren't deleted/etc.
     std::vector<SharedPtr<Thread>> thread_list;
